Add simpler Tau-Leaping solver
The tau-leaping solver previously available could not be used when the
model included Heaviside functions. This is because the implementation
from the org.systemsbiology library computes partial derivatives of the
rate expressions. A new solver has now been added which uses a fixed time
step and hence has no need for the derivative calculations. The new solver
is indicated as "(simpler)" in the GUI, or "tau-leap-simple" from the CLI.
Note that we should still make sure that the old tau-leaping solver flags
up models with non-differentiable functions, as this is currently not
checked.