Allow constant reaction rates
In the previous version of the code, if a reaction's kinetic law
was a simple float value (explicit number or parameter), that
would be interpreted as the constant of a mass-action law rather
than a constant value. Essentially, this did not allow the level
of a species to increase/decrease at a linear rate, but would
result in exponential growth/decay. This was because of how the
rate is interpreted by the systems biology library used (to be
specific, the Reaction class therein), and it affected all solvers
(stochastic and ODE) except for the native Runge-Kutta one. This
commit simply enforces the rate of a reaction to be treatead as an
expression even if it consists of a single number, thus making the
library interpret it correctly.